requirements.txt
google-generativeai
python-dotenv
pillow
google-genai
